Make selendroid-standalone be able to talk to a ipv6 host
When using android emulator, adb binds all their ports to IPv4, if we want to run tests remotely, we need to be able to talk to the server on an IPv6 port as well.
Suppose host A creates the webserver o port B.
We now want to run the tests on host C, but the port B is IPv4, our solution should be on host A to tunnel port B to port D, which is IPv6.
But now, when checking the status of the webserver and making the HTTP requests we want to talk to the IPv6 port.